requirements.txt
cassandra-driver==3.29.1
click==8.1.7
colorama==0.4.6
geomet==0.2.1.post1
jupyter==1.0.0
notebook==7.0.2
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
